+++ elevenlabs_rs/CHANGELOG.md
@@ -17,6 +17,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - `AgentQuery` on `CreateAgent` and `UpdateAgent`
   - `use_tool_ids` Use tool ids instead of tools specs from request payload.   
 - `name` and `access_level` fields to `GetKnowledgeBaseResponse` struct
+- `ListKnowledgeBasesDocs` endpoint
 ### Changed
 - **Breaking**: The `ResponseBody` of `DeleteAgent` now returns a `()` instead of a `StatusResponseBody`
 - **Breaking**: The `used_tools` field name to `tool_ids` of the `PromptConfig` and its type `Option<Vec<UsedTool>>`